initial reorg on evaluation steps
