+++ tests/test_evaluate_model_selection.py
@@ -1,0 +1,13 @@
+import pytest
+
+import wcqtlib.evaluate.model_selection as model_slection
+
+
+@pytest.mark.xfail(reason="Todo")
+def test_ModelSelector():
+    assert False
+
+
+@pytest.mark.xfail(reason="Todo")
+def test_BinarySearchModelSelector():
+    assert False
